backend/requirements.txt
Django>=4.2
djangorestframework
python-decouple
psycopg2-binary
Pillow>=9.0.0

django-extensions
djangorestframework-simplejwt

pytest>=7.0
pytest-django>=4.0
django-filter
rest-framework-roles
django-allauth
dj-rest-auth
requests>=2.0
factory-boy>=3.2.0
pytest-factoryboy
